Massive update: Add ATA implementation for hard drives; Change VFS implementation; Add PCI implementation for managing PCI devices; Add ext2 basic implementation(read, write); Change size_t to u32; Fix bug in realloc; Add test commands to kshell('read', 'write') just for testing ata reads and writes; Remove size_t and uintptr_t from 'types.h' and change NULL to void pointer to 0; Add hdd disk to mastery slave slot in Makefile; Add target to build disk with ext2 file system;

+++ src/heap.h
@@ -22,8 +22,8 @@ void heap_init();
 void *sbrk(u32 increment);
 heap_block *best_fit(u32 size);
 void split_block(heap_block *block, u32 size);
-void *malloc(size_t size);
+void *malloc(u32 size);
 void free(void *ptr);
-void *realloc(void *ptr, size_t size);
+void *realloc(void *ptr, u32 size);
 
 #endif
